+++ index.ts
@@ -1,9 +1,19 @@
-import DiscordJS, {Intents, Message, TextChannel, Permissions, MessageEmbed} from 'discord.js'
+import DiscordJS, {
+  Intents,
+  Message,
+  TextChannel,
+  Permissions,
+  MessageEmbed,
+  MessageActionRow,
+  MessageButton,
+  Interaction,
+  MessageComponentInteraction
+} from 'discord.js'
 import undici from 'undici'
 import jsdom from 'jsdom'
-global.HTMLAnchorElement = new jsdom.JSDOM().window.HTMLAnchorElement
+// global.HTMLAnchorElement = new jsdom.JSDOM().window.HTMLAnchorElement
 import * as db from './database'
-import {sleep, hashCode, msToMinSec, timeString} from './utility'
+import {sleep, hashCode, html2text, text2price, msToMinSec, timeString, diff} from './utility'
 import dotenv from 'dotenv'
 dotenv.config()
 
@@ -36,11 +46,15 @@ class logStack {
 class dataPool {
   link: Array<string>
   title: Array<string>
-  info: Array<number>
+  info: Array<string>
+  hash: Array<number>
+  date: Array<number>
   constructor() {
     this.link = []
     this.title = []
     this.info = []
+    this.hash = []
+    this.date = []
   }
 }
 
@@ -48,17 +62,21 @@ var logger = new logStack()
 var checkerData = new dataPool()
 
 async function init() {
-  const data = await db.getDBShopList()
+  logger.logging('Reading old data.')
+  const data = await db.getShopList()
   for (const shop of data) {
     checkerData.link.push(shop.link)
     checkerData.title.push(shop.title)
     checkerData.info.push(shop.info)
+    checkerData.hash.push(shop.hash)
+    checkerData.date.push(shop.date)
   }
-  logger.logging('Old data is ready to use.')
+  logger.logging(`${data.length} record found`)
 
   const channel = (await client.channels.fetch('938732984973017129')) as TextChannel
   const msg = await channel.messages.fetch({limit: 1}).then(coll => coll.first())
   if (msg === undefined) return
+  if (!msg.content.startsWith('Last updated')) return
   if (client.user !== null && msg.author.id === client.user.id) {
     logger.logging('Auto restart checker.')
     await channel.send(`Auto restart checker.`)
@@ -178,10 +196,116 @@ async function displayChecker(channel: TextChannel, data: dataPool) {
       content += `[${data.title[i]}](${data.link[i]})\n`
     }
     const embed = new MessageEmbed().addFields({name: `${Math.floor(length / 5) * 5 + 1}-${data.link.length}`, value: content})
-    channel.send({
-      embeds: [embed]
-    })
+    channel.send({embeds: [embed]})
   }
+}
+
+async function displayHistory(channel: TextChannel, data: Array<db.Shop>) {
+  var index = 0
+
+  const embed = () =>
+    new MessageEmbed().addFields({
+      name: timeString(data[index].date),
+      value: `[${data[index].title}](${data[index].link})` + '\n' + data[index].info
+    })
+
+  const row = new MessageActionRow()
+    .addComponents(new MessageButton().setCustomId('left').setLabel('<-').setStyle('PRIMARY'))
+    .addComponents(new MessageButton().setCustomId('right').setLabel('->').setStyle('PRIMARY'))
+
+  let botmsg = await channel.send({
+    embeds: [embed()],
+    components: [row]
+  })
+
+  const filter = (butInt: Interaction) => {
+    return butInt.user.id === '472053971406815242'
+  }
+  const collector = channel.createMessageComponentCollector({
+    filter,
+    time: 1000 * 60 * 3
+  })
+  collector.on('collect', async (i: MessageComponentInteraction) => {
+    if (i.customId === 'left') {
+      index ? index-- : (index = data.length - 1)
+    }
+    if (i.customId === 'right') {
+      index < data.length - 1 ? index++ : (index = 0)
+    }
+    botmsg = await botmsg.edit({
+      embeds: [embed()],
+      components: [row]
+    })
+    i.deferUpdate()
+  })
+}
+
+async function displayPrice(channel: TextChannel, data: dataPool) {
+  var index = 0
+
+  const embed = () =>
+    new MessageEmbed().addFields({
+      name: `${index + 1}/${data.link.length}`,
+      value: `[${data.title[index]}](${data.link[index]})` + '\n' + text2price(data.info[index])
+    })
+
+  const row = new MessageActionRow()
+    .addComponents(new MessageButton().setCustomId('left').setLabel('<-').setStyle('PRIMARY'))
+    .addComponents(new MessageButton().setCustomId('right').setLabel('->').setStyle('PRIMARY'))
+
+  let botmsg = await channel.send({
+    embeds: [embed()],
+    components: [row]
+  })
+
+  const filter = (butInt: Interaction) => {
+    return butInt.user.id === '472053971406815242'
+  }
+  const collector = channel.createMessageComponentCollector({
+    filter,
+    time: 1000 * 60 * 3
+  })
+  collector.on('collect', async (i: MessageComponentInteraction) => {
+    if (i.customId === 'left') {
+      index ? index-- : (index = data.link.length - 1)
+    }
+    if (i.customId === 'right') {
+      index < data.link.length - 1 ? index++ : (index = 0)
+    }
+    botmsg = await botmsg.edit({
+      embeds: [embed()],
+      components: [row]
+    })
+    i.deferUpdate()
+  })
+}
+
+async function displayDiff(message: Message) {
+  const filter = (m: Message) => m.author.id === message.author.id
+  const collector = message.channel.createMessageCollector({filter, time: 15000})
+
+  logger.logging(`Try to diff record`)
+  var botMsg = await message.channel.send(`Please enter Link of the shop to diff.`)
+  collector.on('collect', async m => {
+    const data = await db.getShopHistory(m.content, true)
+    if (data.length !== 2) {
+      logger.logging(`Older record not found`)
+      message.channel.send('Older record not found')
+      return
+    }
+    const embed = new MessageEmbed().addFields({
+      name: `info diff`,
+      value: `[${data[0].title}](${data[0].link})` + '\n' + diff(data[0].info, data[1].info)
+    })
+    logger.logging(`successfully diff ${m.content}`)
+    message.channel.send({embeds: [embed]})
+    try {
+      await botMsg.delete()
+      await m.delete()
+    } catch (err) {
+      console.log(err)
+    }
+  })
 }
 
 async function checker() {
@@ -222,31 +346,44 @@ async function checker() {
     if (channel == null) return
     if (old.link.length === 0) {
       displayChecker(channel, current)
+      await db.updateShopList(current.link, current.title, current.info, current.hash)
       return true
     }
+
     var contentList: Array<string> = ['', '', '', '']
     const tag = ['new shop', 'title change', 'info change', 'shop delete']
+
     for (var i = 0; i < current.link.length; i++) {
       const oldIndex = old.link.indexOf(current.link[i])
+
       if (oldIndex === -1) {
         contentList[0] += `[${current.title[i]}](${current.link[i]})\n`
+        await db.updateShopList([current.link[i]], [current.title[i]], [current.info[i]], [current.hash[i]])
         continue
       }
+
       if (current.title[i] != old.title[oldIndex]) {
+        await db.discardShop(current.link[i])
         contentList[1] += `[${old.title[oldIndex]}](${current.link[i]})->\n[${current.title[i]}](${current.link[i]})\n`
+        await db.updateShopList([current.link[i]], [current.title[i]], [current.info[i]], [current.hash[i]])
         continue
       }
+
       if (current.info[i] != old.info[oldIndex]) {
-        contentList[2] += `[${current.title[oldIndex]}](${current.link[i]})\n`
+        await db.discardShop(current.link[i])
+        contentList[2] += `[${current.title[i]}](${current.link[i]})\n`
+        await db.updateShopList([current.link[i]], [current.title[i]], [current.info[i]], [current.hash[i]])
       }
     }
+
     for (var i = 0; i < old.link.length; i++) {
       const newIndex = current.link.indexOf(old.link[i])
       if (newIndex === -1) {
-        await db.deleteDBShopList(old.link[i])
+        await db.discardShop(old.link[i], true)
         contentList[3] += `${old.title[i]}](${old.link[i]})\n`
       }
     }
+
     for (var i = 0; i < contentList.length; i++) {
       if (contentList[i].length > 0) {
         modified = true
@@ -274,23 +411,35 @@ async function checker() {
       const {window} = new jsdom.JSDOM(await body.text())
       const info = window.document.querySelector('#editer_main > div')
       if (info != null) {
+        let images = window.document.querySelectorAll('#editer_main > div > img') as NodeListOf<HTMLImageElement>
+        let str = ''
+        images.forEach(img => {
+          str = str + img.src + '\n'
+        })
         currentData.link.push(link)
         currentData.title.push(window.document.title)
-        currentData.info.push(hashCode(info.innerHTML))
+        currentData.info.push(str + html2text(info.innerHTML))
+        currentData.hash.push(hashCode(info.innerHTML))
+        currentData.date.push(Date.now())
       }
     }
 
     console.log('Checking data.')
     await botMsg.edit(`Checking data.`)
     var modified = await compare(currentData)
-    if (!modified) {
-      await botMsg.edit(`Last updated: ${timeString()}`)
-    } else {
+    var atLast = await channel.messages.fetch({limit: 1}).then(msg => {
+      let lastMsg = msg.first()
+      return lastMsg?.id === botMsg.id
+    })
+    if (modified || !atLast) {
       botMsg.delete()
       botMsg = await channel.send(`Last updated: ${timeString()}`)
+    } else {
+      botMsg.edit(`Last updated: ${timeString()}`)
     }
     checkerData = currentData
-    db.updateDBShopList(checkerData.link, checkerData.title, checkerData.info)
+    //move update to compare()
+    //db.updateDBShopList(checkerData.link, checkerData.title, checkerData.info, checkerData.hash)
 
     //update per 10 min
     for (var i = 0; i < 600; i++) {
@@ -310,7 +459,11 @@ async function dbManger(message: Message, type: string) {
     var botMsg = await message.channel.send(`Please enter Link, Tilte and Info Hash.`)
     collector.on('collect', async m => {
       let data = m.content.split(' ')
-      let result = await db.updateDBShopList([data[0]], [data[1]], [Number(data[2])])
+      if (data.length !== 4) {
+        message.channel.send(`Invalid length.`)
+        return
+      }
+      let result = await db.updateShopList([data[0]], [data[1]], [data[2]], [Number(data[3])])
       if (result[0] === 0) {
         logger.logging(`Record of ${data[0]} successfully created`)
         message.channel.send(`Record of ${data[0]} successfully created`)
@@ -318,17 +471,24 @@ async function dbManger(message: Message, type: string) {
         logger.logging(`Record of ${data[0]} already esixt and update to new values`)
         message.channel.send(`Record of ${data[0]} already esixt and update to new values`)
       }
-      await message.delete()
-      await botMsg.delete()
-      await m.delete()
+      try {
+        await botMsg.delete()
+        await m.delete()
+      } catch (err) {
+        console.log(err)
+      }
     })
   }
   if (type === 'update') {
     logger.logging(`Try to update record`)
-    var botMsg = await message.channel.send(`Please enter the Link of record, new tilte and new Info Hash.`)
+    var botMsg = await message.channel.send(`Please enter the Link of record, new tilte, info and hash.`)
     collector.on('collect', async m => {
       let data = m.content.split(' ')
-      let result = await db.updateDBShopList([data[0]], [data[1]], [Number(data[2])], false)
+      if (data.length !== 4) {
+        message.channel.send(`Invalid length.`)
+        return
+      }
+      let result = await db.updateShopList([data[0]], [data[1]], [data[2]], [Number(data[3])], false)
       if (result[1]) {
         logger.logging(`Record of ${data[0]} successfully updated`)
         message.channel.send(`Record of ${data[0]} successfully updated`)
@@ -336,26 +496,57 @@ async function dbManger(message: Message, type: string) {
         logger.logging(`Record of ${data[0]} not found`)
         message.channel.send(`Record of ${data[0]} not found`)
       }
-      await message.delete()
-      await botMsg.delete()
-      await m.delete()
+      try {
+        await botMsg.delete()
+        await m.delete()
+      } catch (err) {
+        console.log(err)
+      }
     })
   }
   if (type === 'delete') {
     logger.logging(`Try to delete record`)
     var botMsg = await message.channel.send(`Please enter the Link of record`)
     collector.on('collect', async m => {
-      let result = await db.deleteDBShopList(m.content)
-      if (result) {
-        logger.logging(`Record of ${m.content} successfully deleted`)
-        message.channel.send(`Record of ${m.content} successfully deleted`)
+      if (m.content === 'all') {
+        var result = await db.deleteAllShop()
       } else {
-        logger.logging(`Record of ${m.content} not found`)
-        message.channel.send(`Record of ${m.content} not found`)
+        var result = await db.deleteShop(m.content)
       }
-      await message.delete()
-      await botMsg.delete()
-      await m.delete()
+      if (result) {
+        logger.logging(`Successfully deleted ${result} document.`)
+        message.channel.send(`Successfully deleted ${result} document.`)
+      } else {
+        logger.logging(`Record not found`)
+        message.channel.send(`Record not found`)
+      }
+      try {
+        await botMsg.delete()
+        await m.delete()
+      } catch (err) {
+        console.log(err)
+      }
+    })
+  }
+  if (type === 'history') {
+    logger.logging(`Try to get history record`)
+    var botMsg = await message.channel.send(`Please enter the Link of record`)
+    collector.on('collect', async m => {
+      var result = await db.getShopHistory(m.content)
+      if (result.length) {
+        displayHistory(message.channel as TextChannel, result)
+        logger.logging(`Found ${result.length} record`)
+        message.channel.send(`Found ${result.length} record`)
+      } else {
+        logger.logging(`Record not found`)
+        message.channel.send(`Record not found`)
+      }
+      try {
+        await botMsg.delete()
+        await m.delete()
+      } catch (err) {
+        console.log(err)
+      }
     })
   }
 }
@@ -381,9 +572,7 @@ client.on('messageCreate', message => {
 
   if (args[0].slice(2) === 'rangedelete' || args[0].slice(2) === 'rd') {
     if (args.length != 3) {
-      message.channel.send({
-        content: 'Invalid arguments count\nUsage:  !!rangedelete  MessageID1  MessageID2'
-      })
+      message.channel.send('Invalid arguments count\nUsage:  !!rangedelete  MessageID1  MessageID2')
       return
     }
     rangedelete(message)
@@ -394,9 +583,7 @@ client.on('messageCreate', message => {
     let size = 20
     if (args.length == 2) {
       if (isNaN(parseInt(args[1], 10))) {
-        message.channel.send({
-          content: 'Invalid arguments count\nUsage:  !!logs  (size)'
-        })
+        message.channel.send('Invalid arguments count\nUsage:  !!logs  (size)')
         return
       }
       size = parseInt(args[1], 10)
@@ -412,9 +599,7 @@ client.on('messageCreate', message => {
       return
     }
     if (args.length != 2) {
-      message.channel.send({
-        content: 'Invalid arguments count\nUsage:  !!checker  on/off'
-      })
+      message.channel.send('Invalid arguments count\nUsage:  !!checker  on/off/display/price')
       return
     }
     if (args[1] === 'on' && !checkerFlag) {
@@ -427,6 +612,28 @@ client.on('messageCreate', message => {
     if (args[1] === 'display') {
       if (checkerFlag) {
         displayChecker(message.channel as TextChannel, checkerData)
+      } else {
+        message.channel.send({
+          content: 'Checker off.'
+        })
+      }
+    }
+    if (args[1] === 'price') {
+      if (checkerFlag) {
+        displayPrice(message.channel as TextChannel, checkerData)
+      } else {
+        message.channel.send({
+          content: 'Checker off.'
+        })
+      }
+    }
+    if (args[1] === 'diff') {
+      if (checkerFlag) {
+        displayDiff(message)
+      } else {
+        message.channel.send({
+          content: 'Checker off.'
+        })
       }
     }
   }
@@ -436,27 +643,26 @@ client.on('messageCreate', message => {
       return
     }
     if (args.length != 2) {
-      message.channel.send({
-        content: 'Invalid arguments count\nUsage:  !!db  create/update/delete'
-      })
+      message.channel.send('Invalid arguments count\nUsage:  !!db  create/update/delete/history')
       return
     }
-    const options = ['create', 'update', 'delete']
+    const options = ['create', 'update', 'delete', 'history']
     if (options.map(o => args[1] === o).reduce((acc, curr) => acc || curr, false)) {
       dbManger(message, args[1])
     } else {
-      message.channel.send({
-        content: 'Invalid options\nUsage:  !!db  create/update/delete'
-      })
+      message.channel.send('Invalid options\nUsage:  !!db  create/update/delete/history')
       return
     }
   }
 
   if (args[0].slice(2) === 'help') {
-    message.channel.send({
-      content:
-        '**`command list:`**`\n!!rangedelete  MessageID1  MessageID2\n!!logs  (Size)\n!!checker  (on/off/display)\n!!db  (create/update/delete)\n!!help`'
-    })
+    message.channel.send(
+      '**`command list:`**`\n!!rangedelete  MessageID1  MessageID2\n!!logs  (Size)\n!!checker  (on/off/display/price)\n!!db  (create/update/delete/history)\n!!help`'
+    )
+    // message.channel.send({
+    //   content:
+    //     '**`command list:`**`\n!!rangedelete  MessageID1  MessageID2\n!!logs  (Size)\n!!checker  (on/off/display)\n!!db  (create/update/delete)\n!!help`'
+    // })
   }
 })
